refactor(manageuserstable): use new table
asf
