+++ client/app/bundles/course/users/components/buttons/UserManagementButtons.tsx
@@ -2,7 +2,7 @@ import { FC, memo, useState } from 'react';
 import { defineMessages, injectIntl, WrappedComponentProps } from 'react-intl';
 import { toast } from 'react-toastify';
 import equal from 'fast-deep-equal';
-import { CourseUserRowData } from 'types/course/courseUsers';
+import { CourseUserMiniEntity } from 'types/course/courseUsers';
 
 import DeleteButton from 'lib/components/core/buttons/DeleteButton';
 import { COURSE_USER_ROLES } from 'lib/constants/sharedConstants';
@@ -11,7 +11,7 @@ import { useAppDispatch } from 'lib/hooks/store';
 import { deleteUser } from '../../operations';
 
 interface Props extends WrappedComponentProps {
-  user: CourseUserRowData;
+  user: CourseUserMiniEntity;
   disabled?: boolean;
 }
